Move LocalGradientAggregationHelper to common.
